Add test cases for getOrder in OrderInventory
This will ensure that the get order is working as expected when there is invalid input in both the server and method
