+++ code/java_project/componentInventory.json1
@@ -1,27 +1,17 @@
 {
-  "3d1e543a": 45,
-  "e62e6f7d": 43,
-  "2470aaa5": 82,
-  "b39f0014": 37,
-  "dbf74e7c": 94,
-  "ed5d7c98": 17,
-  "2193bd17": 53,
-  "168bebd8": 14,
-  "225cac97": 72,
-  "487a3129": 8,
-  "24b0b5c8": 17,
-  "17faaec7": 20,
-  "10ab25b8": 57,
-  "88b0ff66": 2,
-  "e0b7b726": 93,
-  "9c8a9c9a": 38,
-  "24814c07": 56,
-  "ec7a7562": 42,
-  "644644fa": 66,
-  "ca1edcf8": 18,
-  "0ed555ee": 60,
-  "e66248af": 87,
-  "25025a01": 91,
-  "421a082e": 7,
-  "950a7ebb": 2
+  "6e44bd9c": 1,
+  "ee60dad1": 20,
+  "1d47a55d": 50,
+  "407a7d63": 1,
+  "dd1bb285": 1,
+  "af2ab3f0": 1,
+  "5ea1425a": 1,
+  "6bdb5d0b": 1,
+  "2f09ddce": 1,
+  "e16a4d39": 1,
+  "ffa083cf": 1,
+  "4a37f44d": 1,
+  "bb155838": 1,
+  "9f26f18a": 1,
+  "28471ae9": 5
 }
