+++ code/java_project/src/edu/westga/cs3212/inventory_manager/model/server_impl/ComponentInventory.java
@@ -218,10 +218,6 @@ public class ComponentInventory {
     
     @SuppressWarnings("unchecked")
     private static Map<String, Object> safelyCastToMap(Object object) {
-	if (object instanceof Map<?, ?>) {
-	    return (Map<String, Object>) object;
-	} else {
-	    throw new IllegalArgumentException(Constants.INVALID_RESPONSE_FROM_SERVER);
-	}
+	return (Map<String, Object>) object;
     }
 }
