+++ code/java_project/src/edu/westga/cs3212/inventory_manager/model/server_impl/OrderInventory.java
@@ -197,11 +197,7 @@ public class OrderInventory {
     
     @SuppressWarnings("unchecked")
     private static Map<String, Object> safelyCastToMap(Object object) {
-	if (object instanceof Map<?, ?>) {
-	    return (Map<String, Object>) object;
-	} else {
-	    throw new IllegalArgumentException(Constants.INVALID_RESPONSE_FROM_SERVER);
-	}
+	return (Map<String, Object>) object;
     }
 
 }
