+++ code/java_project/src/edu/westga/cs3212/inventory_manager/test/servercomponentinventory/TestDeleteComponent.java
@@ -31,7 +31,7 @@ class TestDeleteComponent {
 	@Test
 	void testDeleteComponentWithInvalidComponentID() {
 		assertThrows(IllegalArgumentException.class, () -> {
-			ComponentInventory.deleteComponent("invalid");
+			ComponentInventory.deleteComponent("thiscomponentdoesnoteist");
 		});
 	}
