+++ code/java_project/src/edu/westga/cs3212/inventory_manager/test/servercomponentinventory/TestGetQuantity.java
@@ -25,7 +25,7 @@ class TestGetQuantity {
 	@Test
 	void testGetQuantityWithComponentNotInInventory() {
 		assertThrows(IllegalArgumentException.class, () -> {
-			ComponentInventory.getQuantity("123");
+			ComponentInventory.getQuantity("thiscomponentdoesnoteist");
 		});
 	}
