+++ code/java_project/src/edu/westga/cs3212/inventory_manager/test/servercomponentinventory/TestOrderComponent.java
@@ -25,7 +25,7 @@ class TestOrderComponent {
 	@Test 
 	void testOrderComponentWhenCompoentIDIsNotInInventory() {
         assertThrows(IllegalArgumentException.class, () -> {
-            ComponentInventory.orderComponent("123", 0);
+            ComponentInventory.orderComponent("thiscomponentdoesnoteist", 0);
         });
     }
