+++ code/java_project/test.json1
@@ -1,4 +1,4 @@
 {
-  "ca44e17f": 20,
-  "ba3fe3fd": 10
+  "5f20eb7a": 10,
+  "d99799f7": 20
 }
